smartedu-ai/ai-worker/requirements.txt
# SmartEdu AI – AI Worker
openai==1.12.0
httpx==0.27.0
tiktoken==0.5.2
langchain==0.1.6
langchain-openai==0.0.5
pinecone-client==3.0.0
redis==5.0.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.4
google-genai==0.6.0
fastapi==0.109.2
uvicorn==0.27.1
